+++ source/blender/editors/interface/interface_layout.c
@@ -5347,7 +5347,7 @@ static void ui_item_align(uiLayout *litem, short nr)
   }
 }
 
-static void ui_item_flag(uiLayout *litem, uint64_t flag)
+static void ui_item_flag(uiLayout *litem, int flag)
 {
   LISTBASE_FOREACH_BACKWARD (uiItem *, item, &litem->items) {
     if (item->type == ITEM_BUTTON) {
